does refactoring of login and not found pages
