+++ project/package.json
@@ -15,8 +15,7 @@
     "react-redux": "^8.0.4",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
-    "react-toastify": "^9.1.2",
-    "react-usestateref": "^1.0.8"
+    "react-toastify": "^9.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
