+++ project/src/components/filter/filter.tsx
@@ -1,6 +1,7 @@
 import { filterTitleMap } from '../../maps';
 import { NavLink } from 'react-router-dom';
 import cn from 'classnames';
+import './styles.css';
 
 const options = Object.entries(filterTitleMap).map(([optionRoute, optionTitle]) => ({optionRoute, optionTitle}));
