+++ project/src/components/filter/styles.css
@@ -1,0 +1,3 @@
+.locations__item-link span {
+  user-select: none;
+}
