+++ project/src/components/profile/styles.css
@@ -1,0 +1,4 @@
+.header__login,
+.header__signout {
+  user-select: none;
+}
