--- project/src/pages/offer-page/styles.css
@@ -1,3 +1,0 @@
-.property__feature--entire::first-letter {
-  text-transform: uppercase;
-}
